third_party/blink/web_tests/display-lock/lock-before-append/acquire-callback-throws-expected.txt

CONSOLE ERROR: line 23: Uncaught Expected Exception
